requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
scikit-learn>=0.24.0
Pillow>=8.3.0
tqdm>=4.62.0
kagglehub>=0.1.0

# Web Interface Dependencies
Flask>=2.0.0
Flask-WTF>=0.15.0
Werkzeug>=2.0.0
opencv-python>=4.5.0
gunicorn>=20.1.0

# Additional Utilities
tensorboard>=2.8.0
scipy>=1.7.0 